Continued with rule addition, nasty state for the check of valid move
The use of the `test` function results in bad behavior. I check that:
- the score the user keeps is present in the dice
- the dice the user wants to keep corresponds to the score

Two obvious problems:
- With a roll of 3 3 3 4 6, the user can keep 5 2 3 1 1 and have a main
  pleine
- With a roll of 1 2 3 5 6 the user can keep 1 1 4, which doesn't make
  sense

A better approach would be to whitelist the state and check for presence
of dice choice in available dice. For example with a roll of 1 5 5 4 6
the user can pick in either the set 1 5 5 4 6 or 1 1 4 6.